Agentes IA/requirements.txt
crewai
langchain_openai
python-dotenv
